pyproject.toml
[tool.poetry]
name = "object_fit_images"
version = "0.1.0"
description = ""
authors = ["Masaya Abe <49779079+masaya0226@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^9.1.1"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
